build: 优化 Android 构建配置并更新依赖
1.  Android: 将 Gradle 下载源替换为国内镜像以加速构建
2.  Android: 调整 JVM 内存分配，优化构建参数（开启缓存及增量编译，禁用 R8）
3.  Dependencies: 升级 `file_picker` 至 8.0.0，并更新 `pubspec.lock` 中的相关依赖
4.  Dependencies: 调整最低 Flutter SDK 版本要求至 3.38.4
5.  iOS: 适配 `video_player_avfoundation` 插件的类名变更，更新插件注册逻辑

+++ lib/widgets/reader/reader_control_overlay.dart
@@ -250,7 +250,8 @@ class ReaderControlOverlay extends StatelessWidget {
     final maxCh = (totalChapters - 1).toDouble();
     final maxChClamped = maxCh < 0 ? 0.0 : maxCh;
     final cur = (sliderValue >= 0 ? sliderValue : currentChapter.toDouble())
-        .clamp(0, maxChClamped);
+        .clamp(0.0, maxChClamped)
+        .toDouble();
 
     return Row(
       children: [
